content/events/test/test_bug502818.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=502818
-->
<head>
  <title>Test for Bug 502818</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=502818">Mozilla Bug 502818</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 502818 **/

  var count = 0;

  function listener(evt) {
    ++count;
    is(evt.foo, 'bar', "Event.prototype should be accessible from " + evt.type);
  }

  Event.prototype.foo = 'bar';
  window.addEventListener('DOMMouseScroll', listener, false);
  window.addEventListener('MozMousePixelScroll', listener, false);
  window.addEventListener('foobar', listener, false);

  var e = document.createEvent("mousescrollevents");
  e.initMouseScrollEvent("DOMMouseScroll", true, true, window,
                         0, 0, 0, 0, 0, false, false, false, false,
                         0, null, 1);
  window.dispatchEvent(e);

  e.initMouseScrollEvent("MozMousePixelScroll", true, true, window,
                         0, 0, 0, 0, 0, false, false, false, false,
                         0, null, 1);
  window.dispatchEvent(e);

  e.initMouseScrollEvent("foobar", true, true, window,
                         0, 0, 0, 0, 0, false, false, false, false,
                         0, null, 1);
  window.dispatchEvent(e);

  is(count, 3, "Wrong number tests run!");
</script>
</pre>
</body>
</html>
